+++ src/site/markdown/overview/changelog.md
@@ -19,6 +19,9 @@ This is a minor release.
 
 ### API Changes
 
+*   The method `net.sourceforge.pmd.util.StringUtil#htmlEncode(String)` is deprecated.
+    `org.apache.commons.lang3.StringEscapeUtils#escapeHtml4(String)` should be used instead.
+
 ### External Contributions
 
 *   [#368](https://github.com/pmd/pmd/pull/368): \[vf] Adding proper AST support for negation expressions